+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Handle cost, tags, target and location on the Webtools import [#54](https://github.com/rokwire/gateway-building-block/issues/54)
+
 [2.3.0] - 2024-03-18
 ### Added
 - Delete tps events API [#52](https://github.com/rokwire/gateway-building-block/issues/52)